rajout d'un cast pour un retour de ft_strlen dans ft_printf.c

+++ srcs/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: vjovanov <vjovanov@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/02 18:11:46 by vjovanov          #+#    #+#             */
-/*   Updated: 2018/11/21 23:24:44 by vjovanov         ###   ########.fr       */
+/*   Updated: 2019/01/19 17:22:41 by vjovanov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ static int	print_result(t_data *data)
 	else
 		ft_putstr(data->value_format);
 	return (data->identifier == 'c' && data->value[0] == 0 ? i :
-		ft_strlen(data->value_format));
+		(int) ft_strlen(data->value_format));
 }
 
 int			read_percent(const char *format, va_list ap, t_data *data, int *i)
